requirements.txt
django==5.1.14
djangorestframework==3.15.2
django-extensions==4.1